CM-635: Change cron to cron-tasks (#660)
* CM-635: Change cron to cron-tasks

* CM-635: Update cron job

* CM-635: Move publicationState outside filters

* CM-635: Disable updateProtectedAreaStatusCache in cron and fix broken query

* CM-635: Change advisoryDate and expiryDate

* CM-635: Update advisoryDate and expiryDate with lte

* CM-635: Delete updateProtectedAreaStatusCache function

* CM-635: Add populate

---------
