+++ src/cms/config/server.js
@@ -1,4 +1,4 @@
-const cronTasks = require("./cron");
+const cronTasks = require("./cron-tasks");
 
 module.exports = ({ env }) => ({
   host: env("STRAPI_HOST", "0.0.0.0"),
@@ -6,6 +6,7 @@ module.exports = ({ env }) => ({
   url: env("STRAPI_EXTERNAL_URL", "http://localhost:1337"),
   cron: {
     enabled: env.bool("CRON_ENABLED", true),
+    tasks: cronTasks,
   },
   app: {
     keys: env.array('APP_KEYS'),
